Rename SearchContainer to SearchPageContainer
SearchPageContainer is more descriptive than just generic
SearchContainer.
Refs #16.
