+++ app/pages/pages.less
@@ -1,2 +1,2 @@
 @import './resource/resource-page';
-@import './search/search';
+@import './search/search-page';
