+++ app/pages/search/search-page.less
@@ -1,4 +1,4 @@
-.search {
+.search-page {
   h1 {
     color: @red;
   }
