+++ app/pages/search/searchPageSelector.js
