+++ app/state/reducers/index.js
@@ -2,11 +2,11 @@ import { combineReducers } from 'redux';
 import { reducer as formReducer } from 'redux-form';
 
 import { activeRequestsReducer, dataReducer } from 'api/reducers';
-import search from './searchReducer';
+import searchPage from './searchPageReducer';
 
 export default combineReducers({
   activeRequests: activeRequestsReducer,
   data: dataReducer,
   form: formReducer,
-  search,
+  searchPage,
 });
