+++ app/state/reducers/searchPageReducer.js
@@ -1,0 +1,10 @@
+const initialState = {};
+
+function searchPageReducer(state = initialState, action) {
+  switch (action.type) {
+    default:
+      return state;
+  }
+}
+
+export default searchPageReducer;
