--- app/state/reducers/searchReducer.js
@@ -1,10 +1,0 @@
-const initialState = {};
-
-function searchReducer(state = initialState, action) {
-  switch (action.type) {
-    default:
-      return state;
-  }
-}
-
-export default searchReducer;
